implemented correct computation of repetition/definition levels
It is still some kind of a hack, non-leaf writers always sync all state
with parents, not just when leaf writers have non-null data.

Also the code in general is a bit messy, but it works and correctly
determines the levels for the example from the dremel paper.

Added JSONDefinitionTest that uses the new code and documents.json with
the test data from the paper.