rewrite of handling of ColumnWriters and striping, also introduced Tablet
Table is the interface that represents Tablets (collection of columnar
data that represent hierarchical data). Right now we only have dummy
tablet but we will have FileTablet that represents a tablet that is
stored on disk and MemoryTablet that represents a tablet in memory.
Access to both is through the Tablet interface. MemoryTablet will be
used for storing the results of a query and handing it to the
upper-level node.

Tablet has a method getColumnWriter() that gets the column writer for a
specific field in the Schema. FieldStriper now accepts a Tablet and
FieldWriter is a final class that gets passed a ColumnWriter to to the
actual writing to the columnar data.

This should make things a lot more extensible.

Also extended SchemaNode a bit, it now has a parent and a
getQualifiedName() method. Also it has an equals() method now, also a
hashCode() method is provided that uses Objects.hashCode() (guava).