added protobuf description parsing
This uses protostuff librar because the protobuf java library from
google does not provide parsing of .proto files.

Added new class ProtobufSchemaHelper for creating a schema with our
SchemaNodesS from a .proto file and also for writing a schema to a
.proto file.

Changed SchemaNode.toString() again to output in a format that is
compatible with the current .proto format. (No more groups, use nested
messages etc.)

Also added a .proto file for the JSONDefinitionTest that tests the
column striping of the test-data from the dremel paper. This now uses
the .proto loader and does not give the schema "by hand" anymore.