added InMemoryTablet, a Tablet implementation that keeps columns in ram
This is the first implementation of the Tablet interface, the other
being the ondisk implementation that still needs to be done. Also the
ColumnReader is still not defined and InMemoryTablet can therefore not
yet provide column readers.

Also changed the ColumnWriter interface to have a write* method for
every primitive datatype.

Also changed Field and its child classes: Added a writeToColumn() method
that writes the value of the field to the given ColumnWriter using the
correct write* method.