restructured hierarchical data code, new interface RecordStore
RecordStore for now only has a recordIterator() method that gets an
iterator over al fields. But in future it will also have
startNewRecord() method that returns a RecordEncoder, this will then be
used by the record assembly code...

The json implementation had to be changed for this, as well as the
FieldStriper that now accepts a RecordStore instead of a RecordProvider.