massive rewrite of Tablet stuff
Unified handling of in-memory tablet and on-disk tablets by creating a
ColumnReaderImpl and ColumnWriterImpl that operate on InputStream
(OutputStream). OnDiskTablet can now create a reader/writer from file
streams and InMemoryTablets use
ByteArrayIntputStream/ByteArrayOutputStream.

Also split the InMemoryTablet into InMemoryReadonlyTablet and
InMemoryWriteonlyTablet, because one needs only oneway operations with
in-memory tablet and this simplifies synchronization problems with the
in-memory byte-arrays.

Also changed the way IOExceptions are "handled" all the tablet and
writer stuff now simply declares "throws IOException" also FieldStriper
and AssemblyFSM, the code outside of this must handle those exceptions
now.

With this rewrite it should be farely easy to send in-memory tablets
over the network.