restructuring/cleanup of network/transport code
Modified Coordinator, Leaf and CLIClient to have a start() and a stop()
method that do all the connection/cleanup stuff. Also added a shutdown
hook so that netty is correctly shut down on CTRL-C.

Also removed a lot of unhelpful/unnecessary comments/javadoc that were
just cluttering the code.

Renamed Pressluft to SimpleMessage and renamed the pressluft package to
network.messages. Maybe we will add more message types after the
prototype presentation on monday (23. Jan).