first version that actually supports real queries
Modified Coordinator, Leaf and CLIClient, QueryHandler for this and also
introduced helperclass TabletCopier that can copy data from Tablet
(interface) to another Tablet by only using the ColumnReader and
ColumnWriter interfaces.

For now only select * from bla queries are supported and also the query
is hardcoded in CLIClient for now ...

The leafs simply use TabletCopier to copy a tablet to a created
InMemoryWriteonlyTablet that is then sent to the Coordinator.

Also removed PressluftTest from the test folder, renamed it to QueryTest
and put it in the scratch folder becuase it is not an actual unittest.

Also added fill-test-datafir.sh that generates tablets from
Documents.json in the dir specified as argument. This must be done
before running QueryTest or running the system manually.