First commit of massive query / message / network rewrite
Added ANTLR bases "SQL" parser and new classes for representing a parsed
query (Query, SelectClause, WhereClause).

Added new message classes QueryMessage (contains a Query and a query id)
and TabletMassage (contains tablet data as a byte-array and a query id)
and changed Encoder and Decoder to be able to encode/decode them.
SimpleMessage is still retained but the remaining types of messages
should also be converted to the new message scheme.